requirements.txt
openai==1.40.6
opencv-python-headless==4.10.0.84
pillow==10.4.0
typer==0.12.3
